requirements.txt
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
kaggle==1.5.16
pyyaml==6.0.1
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0 